+++ lib/new_relic/distributed_trace/w3c_trace_context.ex
@@ -13,44 +13,44 @@ defmodule NewRelic.DistributedTrace.W3CTraceContext do
          tracestate_header <- Plug.Conn.get_req_header(conn, @w3c_tracestate),
          %TraceParent{} = traceparent <- TraceParent.decode(traceparent_header),
          %TraceState{} = tracestate <- TraceState.decode(tracestate_header) do
-      case TraceState.new_relic(tracestate) do
-        {[], others} ->
-          %Context{
-            source:
-              {:w3c,
-               %{
-                 tracestate: :non_new_relic,
-                 others: others,
-                 sampled: traceparent.flags.sampled,
-                 tracing_vendors: Enum.map(others, & &1.key) |> Enum.join(",")
-               }},
-            parent_id: traceparent.parent_id,
-            span_guid: traceparent.parent_id,
-            trace_id: traceparent.trace_id,
-            trust_key: Collector.AgentRun.trusted_account_key()
-          }
-
-        {[%{value: tracestate}], others} ->
+      case TraceState.restrict_access(tracestate) do
+        {newrelic, others} ->
           %Context{
             source:
               {:w3c,
                %{
                  tracestate: :new_relic,
-                 span_id: tracestate.span_id,
-                 others: others,
                  sampled: traceparent.flags.sampled,
-                 tracing_vendors: Enum.map(others, & &1.key) |> Enum.join(",")
+                 others: others,
+                 span_id: newrelic.span_id,
+                 tracing_vendors: Enum.map(others, & &1.key)
                }},
-            type: tracestate.parent_type,
-            account_id: tracestate.account_id,
-            app_id: tracestate.app_id,
-            parent_id: tracestate.transaction_id,
+            type: newrelic.parent_type,
+            account_id: newrelic.account_id,
+            app_id: newrelic.app_id,
+            parent_id: newrelic.transaction_id,
             span_guid: traceparent.parent_id,
             trace_id: traceparent.trace_id,
-            trust_key: tracestate.trusted_account_key,
-            priority: tracestate.priority,
-            sampled: tracestate.sampled,
-            timestamp: tracestate.timestamp
+            trust_key: newrelic.trusted_account_key,
+            priority: newrelic.priority,
+            sampled: newrelic.sampled,
+            timestamp: newrelic.timestamp
+          }
+
+        others ->
+          %Context{
+            source:
+              {:w3c,
+               %{
+                 tracestate: :non_new_relic,
+                 sampled: traceparent.flags.sampled,
+                 others: others,
+                 tracing_vendors: Enum.map(others, & &1.key)
+               }},
+            parent_id: traceparent.parent_id,
+            span_guid: traceparent.parent_id,
+            trace_id: traceparent.trace_id,
+            trust_key: Collector.AgentRun.trusted_account_key()
           }
       end
     else
